# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IndiaConnect Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/sellers:
    get:
      operationId: listAdminSellers
      tags: [admin]
      summary: List all sellers for admin (with filters)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Sellers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSellerListResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/sellers/{id}/approve:
    post:
      operationId: approveSeller
      tags: [admin]
      summary: Approve a pending seller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seller approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/sellers/{id}/reject:
    post:
      operationId: rejectSeller
      tags: [admin]
      summary: Reject a pending seller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerRejectionInput"
      responses:
        "200":
          description: Seller rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{id}/toggle:
    post:
      operationId: toggleProductActive
      tags: [admin]
      summary: Toggle product active/inactive status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/services/{id}/toggle:
    post:
      operationId: toggleServiceActive
      tags: [admin]
      summary: Toggle service active/inactive status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth schemas ────────────────────────────────────────────────────
...
    Product:
      type: object
      properties:
        id:
          type: integer
        seller_id:
          type: integer
        seller_name:
          type: string
        seller_city:
          type: string
        category_id:
          type: integer
        category_name:
          type: string
        title:
          type: string
        description:
          type: string
        price_min:
          type: ["number", "null"]
        price_max:
          type: ["number", "null"]
        unit:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        is_featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - seller_id
        - seller_name
        - seller_city
        - category_id
        - category_name
        - title
        - description
        - unit
        - images
        - status
        - is_featured
        - tags
        - created_at
...
    Service:
      type: object
      properties:
        id:
          type: integer
        seller_id:
          type: integer
        seller_name:
          type: string
        seller_city:
          type: string
        category_id:
          type: integer
        category_name:
          type: string
        title:
          type: string
        description:
          type: string
        price_min:
          type: ["number", "null"]
        price_max:
          type: ["number", "null"]
        price_unit:
          type: string
          enum: [per_hour, per_day, per_project, fixed, negotiable]
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        is_featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - seller_id
        - seller_name
        - seller_city
        - category_id
        - category_name
        - title
        - description
        - price_unit
        - images
        - status
        - is_featured
        - tags
        - created_at
...
    SellerSummary:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        user_email:
          type: ["string", "null"]
        business_name:
          type: string
        gst_number:
          type: string
        verification_status:
          type: string
          enum: [pending, approved, rejected]
        city:
          type: string
        state:
          type: string
        product_count:
          type: integer
        service_count:
          type: integer
        created_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - user_name
        - business_name
        - gst_number
        - verification_status
        - city
        - state
        - product_count
        - service_count
        - created_at
...
    AdminSellerListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SellerSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - total_pages
    SellerRejectionInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 5

    # ── Dashboard schemas ───────────────────────────────────────────────
...