# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * IndiaConnect Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ServicePriceUnit } from './servicePriceUnit';
import type { ServiceStatus } from './serviceStatus';

export interface Service {
  id: number;
  seller_id: number;
  seller_name: string;
  seller_city: string;
  category_id: number;
  category_name: string;
  title: string;
  description: string;
  /** @nullable */
  price_min?: number | null;
  /** @nullable */
  price_max?: number | null;
  price_unit: ServicePriceUnit;
  images: string[];
  status: ServiceStatus;
  is_featured: boolean;
  tags: string[];
  created_at: Date;
}